+++ Part_2/assignment.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include "curve.h"
 
-int lloyds_assignment(std::vector<real_curve *> & centroids,
+double lloyds_assignment(std::vector<real_curve *> & centroids,
   std::vector<real_curve *> & curves,
   std::vector<std::vector<real_curve *>> & assignment);
