+++ Part_2/update.h
@@ -4,7 +4,7 @@
 #include "curve.h"
 
 void mean_frechet_dist();
-int pam_update(std::vector<real_curve *> & centroids,
-  std::vector<std::vector<real_curve *>> & assignment, int objective);
+double pam_update(std::vector<real_curve *> & centroids,
+  std::vector<std::vector<real_curve *>> & assignment, double objective);
 
 #endif
